v2.00.00
Modified users and profiles so that additional logins can be created
for the web-console.
Modified so application private key is kept in DB instead of filesystem.
Added public SSH key management.
Added auditing of SSH sessions.
Changed to support multiple simultaneous users.
Updated project dependencies.

+++ src/main/java/com/keybox/manage/db/ScriptDB.java
@@ -37,9 +37,10 @@ public class ScriptDB {
     /**
      * returns scripts based on sort order defined
      * @param sortedSet object that defines sort order
+     * @param userId user id
      * @return sorted script list
      */
-    public static SortedSet getScriptSet(SortedSet sortedSet) {
+    public static SortedSet getScriptSet(SortedSet sortedSet, Long userId) {
 
         ArrayList<Script> scriptList = new ArrayList<Script>();
 
@@ -48,12 +49,13 @@ public class ScriptDB {
         if (sortedSet.getOrderByField() != null && !sortedSet.getOrderByField().trim().equals("")) {
             orderBy = "order by " + sortedSet.getOrderByField() + " " + sortedSet.getOrderByDirection();
         }
-        String sql = "select * from scripts " + orderBy;
+        String sql = "select * from scripts where user_id=? " + orderBy;
 
         Connection con = null;
         try {
             con = DBUtils.getConn();
             PreparedStatement stmt = con.prepareStatement(sql);
+            stmt.setLong(1, userId);
             ResultSet rs = stmt.executeQuery();
 
             while (rs.next()) {
@@ -81,15 +83,16 @@ public class ScriptDB {
     /**
      * returns script base on id
      * @param scriptId script id
+     * @param userId user id
      * @return script object
      */
-    public static Script getScript(Long scriptId) {
+    public static Script getScript(Long scriptId, Long userId) {
 
         Script script = null;
         Connection con = null;
         try {
             con = DBUtils.getConn();
-            script = getScript(con, scriptId);
+            script = getScript(con, scriptId, userId);
 
 
         } catch (Exception e) {
@@ -104,14 +107,16 @@ public class ScriptDB {
      * returns script base on id
      * @param con DB connection
      * @param scriptId script id
+     * @param userId user id
      * @return script object
      */
-    public static Script getScript(Connection con, Long scriptId) {
+    public static Script getScript(Connection con, Long scriptId, Long userId) {
 
         Script script = null;
         try {
-            PreparedStatement stmt = con.prepareStatement("select * from  scripts where id=?");
+            PreparedStatement stmt = con.prepareStatement("select * from  scripts where id=? and user_id=?");
             stmt.setLong(1, scriptId);
+            stmt.setLong(2, userId);
             ResultSet rs = stmt.executeQuery();
 
             while (rs.next()) {
@@ -133,16 +138,18 @@ public class ScriptDB {
     /**
      * inserts new script
      * @param script script object
+     * @param userId user id
      */
-    public static void insertScript(Script script) {
+    public static void insertScript(Script script, Long userId) {
 
 
         Connection con = null;
         try {
             con = DBUtils.getConn();
-            PreparedStatement stmt = con.prepareStatement("insert into scripts (display_nm, script) values (?,?)");
+            PreparedStatement stmt = con.prepareStatement("insert into scripts (display_nm, script, user_id) values (?,?,?)");
             stmt.setString(1, script.getDisplayNm());
             stmt.setString(2, script.getScript());
+            stmt.setLong(3, userId);
             stmt.execute();
             DBUtils.closeStmt(stmt);
 
@@ -156,17 +163,19 @@ public class ScriptDB {
     /**
      * updates existing script
      * @param script script object
+     * @param userId user id
      */
-    public static void updateScript(Script script) {
+    public static void updateScript(Script script, Long userId) {
 
 
         Connection con = null;
         try {
             con = DBUtils.getConn();
-            PreparedStatement stmt = con.prepareStatement("update scripts set display_nm=?, script=? where id=?");
+            PreparedStatement stmt = con.prepareStatement("update scripts set display_nm=?, script=? where id=? and user_id=?");
             stmt.setString(1, script.getDisplayNm());
             stmt.setString(2, script.getScript());
             stmt.setLong(3, script.getId());
+            stmt.setLong(4, userId);
             stmt.execute();
             DBUtils.closeStmt(stmt);
 
@@ -180,15 +189,17 @@ public class ScriptDB {
     /**
      * deletes script
      * @param scriptId script id
+     * @param userId user id
      */
-    public static void deleteScript(Long scriptId) {
+    public static void deleteScript(Long scriptId, Long userId) {
 
 
         Connection con = null;
         try {
             con = DBUtils.getConn();
-            PreparedStatement stmt = con.prepareStatement("delete from scripts where id=?");
+            PreparedStatement stmt = con.prepareStatement("delete from scripts where id=? and user_id=?");
             stmt.setLong(1, scriptId);
+            stmt.setLong(2, userId);
             stmt.execute();
             DBUtils.closeStmt(stmt);
